refactor: common module 의 user business dto 패키지 구조 변경
