+++ src/test/java/com/food/app/AuthenticationTest.java
@@ -11,9 +11,9 @@ import com.food.common.user.enumeration.Role;
 import com.food.common.user.repository.AppAccountRepository;
 import com.food.common.user.repository.StoreOwnerRepository;
 import com.food.common.user.repository.UserRepository;
-import com.food.common.user.business.account.impl.response.FoundAppAccount;
-import com.food.common.user.business.user.impl.response.FoundUser;
-import com.food.common.user.business.AccountFindResponse;
+import com.food.common.user.business.dto.response.accountDomain.FoundAppAccount;
+import com.food.common.user.business.dto.response.userDomain.FoundUser;
+import com.food.common.user.business.dto.response.accountFind.AccountFindResponse;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
